Streamlit/requirements.txt
streamlit
mysql.connector
pandas
datetime
